Compute subclassification ATT and ATE

Compute subclassification ATT and ATE

subclassify(df, x, y = "re78", w = "treat", debug = F)

Arguments

df

data.table

x

character vector of discrete covariates

y

outcome

w

treatment

debug

boolean to return strata-specific estimates

Value

list with ATE and ATT estimate in est and strata level averages in table

Examples

data(lalonde.psid); setDT(lalonde.psid)
#> Error in data(lalonde.psid): could not find function "data"
#> Error in setDT(lalonde.psid): Cannot convert 'lalonde.psid' to data.table by reference because binding is locked. It is very likely that 'lalonde.psid' resides within a package (or an environment) that is locked to prevent modifying its variable bindings. Try copying the object to your current environment, ex: var <- copy(var) and then using setDT again.
subclassify(lalonde.psid, x = c("u74", "u75"),
  y = "re78", w = "treat")
#> Error in as.formula(paste0(paste(x, collapse = "+"), "~", w)): could not find function "as.formula"
subclassify(lalonde.psid, x = c("married", "black", "hispanic", "u74", "u75"),
  y = "re78", w = "treat")
#> Error in as.formula(paste0(paste(x, collapse = "+"), "~", w)): could not find function "as.formula"