Abbreviated Regression Summary Generic summaries for lm, glm and mer objects This generic function provides an abbreviated regression output containing the more useful information. Users wanting to see more are advised to use summary()

Abbreviated Regression Summary Generic summaries for lm, glm and mer objects This generic function provides an abbreviated regression output containing the more useful information. Users wanting to see more are advised to use summary()

Arguments

object

An lm, glm or mer object returned from lm(), glm() or lmer() respectively

...

further arguments passed to or from other methods.

Value

returns the same as summary()

See also

Author

Julian Faraway

Examples

data(stackloss)
object <- lm(stack.loss ~ .,stackloss)
sumary(object)
#>              Estimate Std. Error t value  Pr(>|t|)
#> (Intercept) -39.91967   11.89600 -3.3557   0.00375
#> Air.Flow      0.71564    0.13486  5.3066 5.799e-05
#> Water.Temp    1.29529    0.36802  3.5196   0.00263
#> Acid.Conc.   -0.15212    0.15629 -0.9733   0.34405
#> 
#> n = 21, p = 4, Residual SE = 3.24336, R-Squared = 0.91